Enemy AI and Trap
Made one enemy type that follows four set points on the map (can have any amount of waypoints if needed). If the enemy detects the player, the enemy will chase the player. While chasing, if the enemy doesn't detect the player for a set amount of time, the enemy will return to following the set points. If the enemy collides with a trap, the enemy will be slowed down for a set amount of time.

[SerializeFIeld] variables:

Waypoints
ChaseTImerMax: Time that is set to chase player when the player is detected
WalkSpeed: movement speed while patroling
ChaseSpeed: Movement speed while chasing the player
TrapSpeed: Movement speed while trapped
TrapTimerMax: Set time that movement speed is while trapped