Arrow Prefab
Added arrow object that moves on its own. When it collides with an enemy, it stops moving and becomes a child of the enemy body object. It will then disappear after 6 seconds (can be changed in code). While on the enemy, it will slow the enemy down the same way a trap will.