New Enemy Type, Added Models, and Animations
Coded Enemies to prioritize closest followers if detected. Requires followers to have layer "Follower". Added Snowman model for Still Movement Enemies and Golem model for Patrol Enemies. Added animations to Golems that are dependent on movement speed. (If you want to change the movement speeds of a golem, you'll have to also change the corresponding animation speeds. Created Prefabs of both enemy types. Each time you use a prefab, you'll have to assign all followers so they can be detected, the player so it can be detected, and waypoints (Empty gameObjects) for Golems.